+++ be/src/main/java/com/issuetrackermax/controller/filter/dto/response/FilterResponse.java
@@ -1,0 +1,27 @@
+package com.issuetrackermax.controller.filter.dto.response;
+
+import java.util.List;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class FilterResponse {
+	private Long labelCount;
+	private Long mileStoneCount;
+	private Long openIssueCount;
+	private Long closedIssueCount;
+	private List<IssueResponse> issues;
+
+	@Builder
+	public FilterResponse(Long labelCount, Long mileStoneCount, Long openIssueCount, Long closedIssueCount,
+		List<IssueResponse> issues) {
+		this.labelCount = labelCount;
+		this.mileStoneCount = mileStoneCount;
+		this.openIssueCount = openIssueCount;
+		this.closedIssueCount = closedIssueCount;
+		this.issues = issues;
+	}
+}
