+++ be/src/test/java/com/issuetrackermax/service/jwt/JwtServiceTest.java
@@ -7,6 +7,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.issuetrackermax.common.exception.InCorrectPasswordException;
+import com.issuetrackermax.common.exception.InvalidLoginIdException;
+import com.issuetrackermax.common.exception.InvalidPasswordException;
 import com.issuetrackermax.controller.member.dto.request.SignUpRequest;
 import com.issuetrackermax.domain.IntegrationTestSupport;
 import com.issuetrackermax.domain.jwt.entity.Jwt;
@@ -39,12 +42,12 @@ class JwtServiceTest extends IntegrationTestSupport {
 		SignUpRequest signUpRequest = SignUpRequest.builder()
 			.loginId("June@codesquad.co.kr")
 			.nickName("June")
-			.password("1234")
+			.password("12345678")
 			.build();
 		memberService.registerMember(signUpRequest);
 
 		// when
-		Jwt jwt = jwtService.login("June@codesquad.co.kr", "1234");
+		Jwt jwt = jwtService.login("June@codesquad.co.kr", "12345678");
 		String accessToken = jwt.getAccessToken();
 		String refreshToken = jwt.getRefreshToken();
 
@@ -56,18 +59,55 @@ class JwtServiceTest extends IntegrationTestSupport {
 
 	@DisplayName("패스워드가 일치하지 않으면 오류를 일으킨다.")
 	@Test
-	void login2() throws Exception {
+	void incoreectPasswordException() throws Exception {
+		// given
+		SignUpRequest signUpRequest = SignUpRequest.builder()
+			.loginId("June@codesquad.co.kr")
+			.nickName("June")
+			.password("12345678")
+			.build();
+		// when
+		memberService.registerMember(signUpRequest);
+
+		// then
+		assertThatThrownBy(() -> jwtService.login("June@codesquad.co.kr", "12345679")).isInstanceOf(
+			InCorrectPasswordException.class);
+	}
+
+	@DisplayName("패스워드가 8글자 이하면 오류를 일으킨다.")
+	@Test
+	void invalidPasswordException() throws Exception {
 		// given
 		SignUpRequest signUpRequest = SignUpRequest.builder()
 			.loginId("June@codesquad.co.kr")
 			.nickName("June")
 			.password("1234")
 			.build();
-		// when
-		memberService.registerMember(signUpRequest);
+		// when & then
+		assertThatThrownBy(() -> memberService.registerMember(signUpRequest)).isInstanceOf(
+			InvalidPasswordException.class);
 
-		// then
-		assertThatThrownBy(() -> jwtService.login("June@codesquad.co.kr", "123")).isInstanceOf(Exception.class);
+	}
+
+	@DisplayName("이미 같은 아이디가 존재하면 오류를 일으킨다.")
+	@Test
+	void invalidLoginException() throws Exception {
+		// given
+		SignUpRequest signUpRequest = SignUpRequest.builder()
+			.loginId("June@codesquad.co.kr")
+			.nickName("June")
+			.password("12345678")
+			.build();
+		memberService.registerMember(signUpRequest);
+		SignUpRequest invalidSignUpRequest = SignUpRequest.builder()
+			.loginId("June@codesquad.co.kr")
+			.nickName("June")
+			.password("12345678")
+			.build();
+
+		// when & then
+		assertThatThrownBy(() -> memberService.registerMember(invalidSignUpRequest)).isInstanceOf(
+			InvalidLoginIdException.class);
 	}
 
 	@DisplayName("accessToken의 만료기간이 지나 refreshToken을 통해 accessToken을 재발행받는다.")
@@ -77,10 +117,10 @@ class JwtServiceTest extends IntegrationTestSupport {
 		SignUpRequest signUpRequest = SignUpRequest.builder()
 			.loginId("June@codesquad.co.kr")
 			.nickName("June")
-			.password("1234")
+			.password("12345678")
 			.build();
 		memberService.registerMember(signUpRequest);
-		Jwt jwt = jwtService.login("June@codesquad.co.kr", "1234");
+		Jwt jwt = jwtService.login("June@codesquad.co.kr", "12345678");
 		// when
 		Jwt newJwt = jwtService.reissueAccessToken(jwt.getRefreshToken());
 		String accessToken = newJwt.getAccessToken();
